# public/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Reservas medicas",
        "description": "API para gestionar las reservas medicas",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Prefijo comun de todas las rutas del router"
        }
    ],
    "paths": {
...
        "/profesionales": {
            "get": {
                "tags": [
                    "Profesionales"
                ],
                "summary": "Listado de los profesionales. Opcionalmente se pueden pasar filtros como query params (nombre, apellido, email, telefono, profesion, ciudad, direccion)",
                "operationId": "7167299be780cdfbff07e37265a36f17",
                "parameters": [
                    {
                        "name": "nombre",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "apellido",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "telefono",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "profesion",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ciudad",
                        "in": "query",
                        "description": "Filtra por ciudad del consultorio",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "direccion",
                        "in": "query",
                        "description": "Filtra por direccion del consultorio",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de profesionales (opcionalmente filtrada)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/RespuestaProfesional"
                                            }
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 42
                                        },
                                        "page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "limit": {
                                            "type": "integer",
                                            "example": 10
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Filtro no permitido o valor invalido",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "ERROR": "Filtro no permitido"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/profesionales/{id}": {
            "get": {
                "tags": [
                    "Profesionales"
                ],
                "summary": "Obtener un profesional por su id",
                "operationId": "fe797f76ed23cb1f879d3ee1d6237c8e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Existe un profesional con ese ID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RespuestaProfesional"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "ID invalido",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "ERROR": "El id ingresado no es valido"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se hallo un profesional con ese id",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "ERROR": "No se encontro un profesional con ese id"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Profesionales"
                ],
                "summary": "Baja de un profesional",
                "operationId": "666305152d486c29d9cee827246d46da",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "motivo"
                                ],
                                "properties": {
                                    "motivo": {
                                        "description": "Motivo de la baja del profesional",
                                        "type": "string",
                                        "example": "El profesional ha estado inactivo por mas de 6 meses"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Profesional dado de baja",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Solicitud erronea: ID invalido | no hay motivo de baja | motivo muy largo",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "ERROR": "ID Invalido"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Profesional no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "ERROR": "No se encontro un profesional a eliminar con ese id"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Profesionales"
                ],
                "summary": "Actualizar datos del profesional",
                "operationId": "427c8dbb7f162bfaab851524479bc354",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ActualizarProfesionalRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profesional actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RespuestaProfesional"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Solicitud erronea: Campos invalidos | ID invalido",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "ERROR": "JSON invalido"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Profesional no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "ERROR": "No se encontro un profesional con ese id"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflicto: Usuario ya existente con esos datos!",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "ERROR": "Ya existe un usuario con ese email y/o telefono ingresado/s"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/profesionales/registrar": {
            "post": {
                "tags": [
                    "Profesionales"
                ],
                "summary": "Registrar un profesional",
                "operationId": "cbb6535cbba601ef6d80bebbf9e8dfdb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CrearProfesionalRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Profesional registrado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RespuestaProfesional"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Solicitud erronea: Campos invalidos | formato de contraseña incorrecto",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "ERROR": "La contraseña debe tener un caracter especial!"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflicto: Usuario existente!",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "ERROR": "Ya existe un usuario registrado con ese email y/o telefono ingresado/s"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ActualizarProfesionalRequest": {
                "properties": {
                    "nombre": {
                        "type": "string",
                        "example": "Juan",
                        "nullable": true
                    },
                    "apellido": {
                        "type": "string",
                        "example": "Perez",
                        "nullable": true
                    },
                    "profesion": {
                        "type": "string",
                        "example": "Cardiologo",
                        "nullable": true
                    },
                    "email": {
                        "type": "string",
                        "example": "juanPerez@gmail.com",
                        "nullable": true
                    },
                    "telefono": {
                        "type": "string",
                        "example": "123456789",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "CrearProfesionalRequest": {
                "required": [
                    "nombre",
                    "apellido",
                    "profesion",
                    "email",
                    "telefono",
                    "password"
                ],
                "properties": {
                    "nombre": {
                        "type": "string",
                        "example": "Roberto"
                    },
                    "apellido": {
                        "type": "string",
                        "example": "Falcao"
                    },
                    "profesion": {
                        "type": "string",
                        "example": "Odontologo"
                    },
                    "email": {
                        "type": "string",
                        "example": "roberFalcao@outlook.com",
                        "nullable": true
                    },
                    "telefono": {
                        "type": "string",
                        "example": "0984728910",
                        "nullable": true
                    },
                    "password": {
                        "type": "string",
                        "example": "Rober13#"
                    }
                },
                "type": "object"
            },
            "RespuestaProfesional": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 12
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Lucero Gimenez"
                    },
                    "profesion": {
                        "type": "string",
                        "example": "Pediatra"
                    },
                    "email": {
                        "type": "string",
                        "example": "luceroG@outlook.com",
                        "nullable": true
                    },
                    "telefono": {
                        "type": "string",
                        "example": "347586901",
                        "nullable": true
                    },
                    "activo": {
                        "type": "boolean",
                        "example": "true"
                    },
                    "motivo_baja": {
                        "type": "string",
                        "example": "Cambio de hospital",
                        "nullable": true
                    },
                    "fecha_baja": {
                        "type": "string",
                        "example": "2026-03-1 08:00:13",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingresa tu token JWT aquí para autenticarte",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}